+++ my-app/src/app/model/single-wine.ts
